# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"NewsResponse": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"articles": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "news-app-api",
		"version": "0.0.1",
		"description": "News App API for OpenWeekApps project",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Luis Martinez"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/news": {
			"get": {
				"operationId": "GetNews",
				"responses": {
					"200": {
						"description": "News obtained successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"articles": [
												{
													"_id": "633c92782c95cc70cec4c1da",
													"source": {
														"id": null,
														"name": "Microsoft.com"
													},
													"author": null,
													"title": "Planetary Computer: una apuesta de Microsoft para un futuro sostenible - Microsoft News Center Latinoamérica",
													"description": "Ciudad de México – Planetary Computer de Microsoft es una plataforma que combina un catálogo de múltiples petabytes de datos de monitoreo ambiental a nivel global. Cuenta con aplicaciones int…",
													"url": "https://news.microsoft.com/es-xl/planetary-computer-una-apuesta-de-microsoft-para-un-futuro-sostenible/",
													"urlToImage": "https://news.microsoft.com/wp-content/uploads/prod/sites/41/2022/10/Planetary-Computer-1.png",
													"publishedAt": "2022-10-04T18:15:30Z",
													"content": "Ciudad de México – Planetary Computer de Microsoft es una plataforma que combina un catálogo de múltiples petabytes de datos de monitoreo ambiental a nivel global. Cuenta con aplicaciones intuitivas … [+3835 chars]",
													"category": "business"
												},
												{
													"_id": "633c92952c95cc70cec4c21f",
													"source": {
														"id": null,
														"name": "News-Medical.Net"
													},
													"author": null,
													"title": "McLean Hospital-affiliated clinicians launch virtual treatment program for anxiety and OCD in children, teens - News-Medical.Net",
													"description": "To tackle the growing mental health crisis facing children and adolescents and the difficulties families face accessing treatments for disorders such as anxiety and obsessive compulsive disorder (OCD), two McLean Hospital-affiliated clinicians have launched a…",
													"url": "https://www.news-medical.net/news/20221004/McLean-Hospital-affiliated-clinicians-launch-virtual-treatment-program-for-anxiety-and-OCD-in-children-teens.aspx",
													"urlToImage": "https://d2jx2rerrg6sh3.cloudfront.net/image-handler/picture/2016/3/Children_playing_sunset_-_Zurijeta_8c5bdac77e44431bb1bfec67b9c87208-620x480.jpg",
													"publishedAt": "2022-10-04T18:08:00Z",
													"content": "To tackle the growing mental health crisis facing children and adolescents and the difficulties families face accessing treatments for disorders such as anxiety and obsessive compulsive disorder (OCD… [+5389 chars]",
													"category": "health"
												},
												{
													"_id": "633c92812c95cc70cec4c1f1",
													"source": {
														"id": null,
														"name": "YouTube"
													},
													"author": null,
													"title": "Kaley Cuoco Says \"I Only Had Eyes for\" 'Big Bang Theory' Costar Johnny Galecki | PEOPLE - People",
													"description": "Kaley Cuoco has no shame about experiencing like at first sight with Johnny Galecki.Cuoco, 36, and Galecki, 47, starred as onscreen love interests Penny and ...",
													"url": "https://www.youtube.com/watch?v=YryWClqH22c",
													"urlToImage": "https://i.ytimg.com/vi/YryWClqH22c/hqdefault.jpg",
													"publishedAt": "2022-10-04T18:09:20Z",
													"content": null,
													"category": "entertainment"
												}
											]
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Something went wrong",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"message": "Something went wrong"
										}
									}
								}
							}
						}
					}
				},
				"description": "Endpoint to get News",
				"tags": [
					"News Controller"
				],
				"security": [],
				"parameters": [
					{
						"description": "Category",
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Search by",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}